Include generic date column name in all time series responses
